add build segregation for windows specific tests
